requirements.txt
selenium==4.21.0
pytest==8.2.1
allure-pytest==2.13.5
